GeometryExperiments

Documentation for GeometryExperiments.

(work in progress)

GeometryExperiments.BezierCurve

GeometryExperiments.BezierEvalMethod

GeometryExperiments.EdgeIterator

GeometryExperiments.Ellipsoid

GeometryExperiments.FixedDegree

GeometryExperiments.GranularVector

GeometryExperiments.Horner

GeometryExperiments.HyperSphere

GeometryExperiments.Mesh

GeometryExperiments.Patch

GeometryExperiments.Projection

GeometryExperiments.Rotation

GeometryExperiments.VertexMesh

Base.allunique

GeometryExperiments.ensure_cyclic_edges!

GeometryExperiments.ismanifold

GeometryExperiments.projection

GeometryExperiments.projection_parameter

GeometryExperiments.BezierCurve — Method

Apply Horner's method on the monomial representation of the Bézier curve B = ∑ᵢ aᵢtⁱ with i ∈ [0, n], n the degree of the curve, aᵢ = binomial(n, i) * pᵢ * t̄ⁿ⁻ⁱ and t̄ = (1 - t). Horner's rule recursively reconstructs B from a sequence bᵢ with bₙ = aₙ and bᵢ₋₁ = aᵢ₋₁ + bᵢ * t until b₀ = B.

GeometryExperiments.BezierEvalMethod — Type

Evaluation method used to obtain a point along a Bézier curve from a parametric expression.

GeometryExperiments.EdgeIterator — Type

Iterator on a mesh face which returns a tuple (prev, next, edge, swapped) where prev and next are the previous and next vertices in the edge cycle, edge the corresponding undirected edge and swapped a boolean value indicating whether the vertices of the undirected edge were swapped to yield prev and next.

The face must be a complete cycle of connected edges. Edges are not required to be connected exactly at dst => src points; since they are undirected, the last next vertex must simply be included as one of the two endpoints of the next edge in the face.

GeometryExperiments.Ellipsoid — Type

Ellipsoid(semiaxes)

An ellipsoid with semi-axes semiaxes. Is equivalent to a scaled HyperSphere.

GeometryExperiments.FixedDegree — Type

Fast evaluation in the case of a fixed small number N of control points.

GeometryExperiments.GranularVector — Type

A vector type which can have items deleted without modifying indices of other values.

Features include:

No hashing is ever performed to access its contents.

The underlying buffer is contiguous in memory.

No allocations occur when retrieving elements by indices from that container or when iterating on the whole container.

Elements can be deleted from the container without having to worry about indices of other elements being shifted.

It is furthermore optimized to reuse indices of previously deleted elements when the number of deleted elements is greater than the number of live elements in the collection. For this reason, push! is O(n/2) in the worst case scenario (which should very rarely happen).

Note that buffer slots are never deleted by purpose, except with empty!. Therefore, if the container grows then shrinks logically, memory won't be freed automatically. In this case, you must empty! the container manually or build another one.

GeometryExperiments.Horner — Type

Approximate evaluation using Horner's method. Recommended for a large number of control points, if you can afford a precision loss. See https://en.wikipedia.org/wiki/Horner%27s_method.

GeometryExperiments.HyperSphere — Type

Hypersphere centered around the origin

GeometryExperiments.Mesh — Type

General representation of a two-dimensional mesh embedded in an arbitrary space.

The associated surface needs not be manifold; there can be dangling edges, lone vertices, and faces linked only by a single vertex.

GeometryExperiments.Patch — Type

Patch made by "gluing" curves together.

If compact is set to true, each curve starts from the point of the last curve, assumed to be on the curve.

GeometryExperiments.Projection — Type

Projection of an object of type O onto the first N dimensions.

GeometryExperiments.Rotation — Type

Rotation(Quaternion(1., 0., 0., 0.))

Rotation around the origin implemented from a rotation type R. R can be any kind of rotation which implements a method for rotate(x, r::R) where {R<:RotationType}. In 3D, this can be Quaternion or Euler, for example.

GeometryExperiments.VertexMesh — Type

Mesh represented with vertex and index data.

Base.allunique — Method

Check whether the mesh has duplicate elements.

GeometryExperiments.ensure_cyclic_edges! — Method

Make sure that edges specified in the face form a cycle.

If they don't form a cycle originally, then edges will be reordered so that a cycle can be formed with all edges; otherwise, an error will be thrown.

GeometryExperiments.ismanifold — Method

Return whether the mesh represents the boundary of a 3-dimensional volume, i.e. it is homogeneously made of connected faces and there is no boundary (every edge is attached to exactly two faces).

GeometryExperiments.projection — Method

projection(object, x) -> x′

Project x onto object, and return the resulting point x′.

GeometryExperiments.projection_parameter — Function

projection_parameter(parametric, x) -> t

Project x onto parametric, and return the corresponding value t in parametric's parameter space.